Added error messages to product views for none superusers

+++ products/views.py
@@ -62,6 +62,12 @@ def product_description(request, product_id):
 
 @login_required
 def add_product(request):
+    """ Add a new product to the store """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'That function is reserved for store admins')
+        return redirect(reverse,('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -82,7 +88,11 @@ def add_product(request):
 
 @login_required
 def edit_product(request, product_id):
+    """ Edit existing products """
 
+    if not request.user.is_superuser:
+        messages.error(request, 'That function is reserved for store admins')
+        return redirect(reverse,('home'))
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -106,7 +116,10 @@ def edit_product(request, product_id):
 
 @login_required
 def delete_product(request, product_id):
-
+    """ Delete items from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'That function is reserved for store admins')
+        return redirect(reverse,('home'))
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted')
